witInfiniteScroll refactored into useInfiniteScroll hook
